+++ src/reducers/administratorPageReducer.js
@@ -1,11 +1,17 @@
 import { combineReducers } from "redux"
 import { ADD_ADMIN_FORM_ERROR, ADD_ADMIN_FORM_SUCCESS, ADMINISTRATOR_SIDEBAR_COLLAPSED_TOGGLED, LIST_ADMINS_SUCCESS, PERIODS_ADD_PERIOD, PERIODS_DELETE_PERIOD, PERIODS_RESET, PERIODS_RETRIEVED, PERIODS_UPDATE_FIELD, PERIODS_UPLOAD_ERROR, PERIODS_UPLOAD_SUCCESS, ROOMS_ADD_ROOM, ROOMS_DELETE_ROOM, ROOMS_RESET, ROOMS_RETRIEVED, ROOMS_UPDATE_FIELD, ROOMS_UPLOAD_ERROR, ROOMS_UPLOAD_SUCCESS, UPLOAD_GROUPS_MEETINGS_MEMBERS_FORM_ERROR, UPLOAD_GROUPS_MEETINGS_MEMBERS_FORM_SUCCESS, UPLOAD_USERS_FORM_ERROR, UPLOAD_USERS_FORM_SUCCESS, SET_ADD_COURSE_ERROR, SET_ADD_USER_ERROR } from "../actions/administratorActions"
 
+/**
+ * Error message and success message
+ */
 const initialState = {
   error: null,
   message: null,
 }
 
+/**
+ * Reducer for messages related to uploading the users file
+ */
 const usersReducer = (state = initialState, action) => {
   switch (action.type) {
     case UPLOAD_USERS_FORM_SUCCESS:
@@ -25,6 +31,9 @@ const usersReducer = (state = initialState, action) => {
   }
 }
 
+/**
+ * Reducer for messages related to uploading the various courses file
+ */
 const coursesReducer = (state = initialState, action) => {
   switch (action.type) {
     case UPLOAD_GROUPS_MEETINGS_MEMBERS_FORM_SUCCESS:
@@ -44,6 +53,9 @@ const coursesReducer = (state = initialState, action) => {
   }
 }
 
+/**
+ * Reducer for editting the list of periods
+ */
 const periodsReducer = (state = {
   default: null,
   current: null,
@@ -51,6 +63,7 @@ const periodsReducer = (state = {
   message: null,
 }, action) => {
   switch (action.type) {
+    // when all periods are retrieved from Firebase
     case PERIODS_RETRIEVED:
       return {
         ...state,
@@ -80,6 +93,7 @@ const periodsReducer = (state = {
       let clone = Object.assign({}, state)
       delete clone.current[action.period] 
       return clone
+    // reset to the last save
     case PERIODS_RESET:
       return {
         ...state,
@@ -103,6 +117,9 @@ const periodsReducer = (state = {
   }
 }
 
+/**
+ * Reducer for editting the list of rooms
+ */
 const roomsReducer = (state = {
   default: null,
   current: null,
@@ -110,6 +127,7 @@ const roomsReducer = (state = {
   message: null,
 }, action) => {
   switch (action.type) {
+    // data retrieved from Firebase
     case ROOMS_RETRIEVED:
       return {
         ...state,
@@ -139,6 +157,7 @@ const roomsReducer = (state = {
       let clone = Object.assign({}, state)
       delete clone.current[action.room] 
       return clone
+    // reset to the last save
     case ROOMS_RESET:
       return {
         ...state,
@@ -162,6 +181,9 @@ const roomsReducer = (state = {
   }
 }
 
+/**
+ * Reducer for data related to the annual board subpage
+ */
 const annualBoardReducer = combineReducers({
   users: usersReducer,
   courses: coursesReducer,
@@ -169,6 +191,9 @@ const annualBoardReducer = combineReducers({
   rooms: roomsReducer,
 })
 
+/**
+ * Reducer for managing the visibility of the sidebar
+ */
 const sidebarCollapsedReducer = (state=true, action) => {
   switch (action.type) {
     case ADMINISTRATOR_SIDEBAR_COLLAPSED_TOGGLED:
@@ -178,6 +203,9 @@ const sidebarCollapsedReducer = (state=true, action) => {
   }
 }
 
+/**
+ * Redcuer for managing the add admin form
+ */
 const adminReducer = (state=initialState, action) => {
   switch (action.type) {
     case ADD_ADMIN_FORM_SUCCESS:
@@ -195,6 +223,9 @@ const adminReducer = (state=initialState, action) => {
   }
 }
 
+/**
+ * Reducer for managing the all current admins display
+ */
 const adminsReducer = (state=[], action) => {
   switch (action.type) {
     case LIST_ADMINS_SUCCESS:
@@ -204,12 +235,17 @@ const adminsReducer = (state=[], action) => {
   }
 }
 
+/** 
+ * Reducer for the Dashboard subpage
+ */
 const dashboardReducer = combineReducers({
   admin: adminReducer,
   admins: adminsReducer,
 })
 
-
+/**
+ * Reducer for the Add Course subpage, with a potential error message
+ */
 const addCourseReducer = (state=null, action) => {
   switch (action.type) {
     case SET_ADD_COURSE_ERROR:
@@ -219,6 +255,9 @@ const addCourseReducer = (state=null, action) => {
   }
 }
 
+/**
+ * Reducer for the Add User subpage, with a potential error message
+ */
 const addUserReducer = (state=null, action) => {
   switch (action.type) {
     case SET_ADD_USER_ERROR:
@@ -228,6 +267,9 @@ const addUserReducer = (state=null, action) => {
   }
 }
 
+/**
+ * Reducer for the Administrator Page
+ */
 const administratorPageReducer = combineReducers({
   annualBoard: annualBoardReducer,
   sidebarCollapsed: sidebarCollapsedReducer,
