+++ src/reducers/index.js
@@ -11,18 +11,21 @@ import meetingPageReducer from './meetingPageReducer'
 import notificationsReducer from './notificationsReducer'
 import profilePageReducer from './profilePageReducer'
 
+/**
+ * Root reducer, combining everything to form the store
+ */
 const createRootReducer = (history) => combineReducers({
-  form: formReducer,
-  firebase: firebaseReducer,
-  firestore: firestoreReducer,
-  location: locationReducer,
-  notifications: notificationsReducer,
+  form: formReducer, // Redux-Form
+  firebase: firebaseReducer, // React-Redux-Firebase
+  firestore: firestoreReducer, // React-Redux-Firebase
+  location: locationReducer, // for the router
+  notifications: notificationsReducer, // custom notifications
   loginPage: loginPageReducer,
   administratorPage: administratorPageReducer,
   bookPage: bookPageReducer,
   meetingPage: meetingPageReducer,
   profilePage: profilePageReducer,
-  router: connectRouter(history),
+  router: connectRouter(history), // for the router
 })
