+++ src/reducers/notificationsReducer.js
@@ -1,6 +1,9 @@
 import { combineReducers } from "redux"
 import { NOTIFICATION_CLOSED, NOTIFICATION_SET } from "../actions/notificationsActions"
 
+/**
+ * Reducer for which current notification is open
+ */
 const currentReducer = (state=null, action) => {
   switch (action.type) {
     case NOTIFICATION_SET:
@@ -12,6 +15,9 @@ const currentReducer = (state=null, action) => {
  } 
 }
 
+/**
+ * Reducer for providing optional data to the currently open notification
+ */
 const dataReducer = (state={}, action) => {
   switch (action.type) {
     case NOTIFICATION_SET:
@@ -21,6 +27,9 @@ const dataReducer = (state={}, action) => {
   }
 }
 
+/**
+ * Reducer for all notifications
+ */
 const notificationsReducer = combineReducers({
   current: currentReducer,
   data: dataReducer,
