+++ src/store/configureStore.js
@@ -1,3 +1,6 @@
+/**
+ * Exports the appropriate Store based on environment
+ */
 if (process.env.NODE_ENV === 'production') {
   module.exports = require('./configureStore.prod')
 } else {
