+++ src/store/history.js
@@ -1,5 +1,8 @@
 import { createBrowserHistory } from 'history' 
 
+/**
+ * Exports a History object for the router
+ */
 const history = createBrowserHistory()
 
 export default history
