+++ src/stylesheets/abstracts/_variables.scss
@@ -10,6 +10,7 @@ $text-font-stack: 'acumin-pro', sans-serif !default;
 /// @type List
 $code-font-stack: 'Courier New', 'DejaVu Sans Mono', 'Bitstream Vera Sans Mono', 'Monaco', monospace !default;
 
+/// Color palette
 $color-primary: #3a4f56;
 $color-secondary: #d32f2f;
 $color-tertiary: #eef5f7;
