+++ src/stylesheets/base/_typography.scss
@@ -1,6 +1,10 @@
 /**
  * Basic typography style for copy text
  */
+
+/**
+ * Sets text color and basic font
+ */
 body {
   color: $text-color;
   font-family: $text-font-stack;
