+++ src/stylesheets/components/_box.scss
@@ -1,3 +1,7 @@
+/**
+ * Individaul box container inside a schedule
+ */
+ 
 .box {
   border: 1px solid $dark-grey;
   padding: 8px 10px 10px;
