+++ src/stylesheets/components/_divider.scss
@@ -1,3 +1,6 @@
+/**
+ * Horizontal line, dividing paragraphs
+ */
 .divider {
   border-color: rgba($mid-grey, .7);
   border-style: inset;
