+++ src/stylesheets/components/_footer.scss
@@ -1,3 +1,7 @@
+/**
+ * Footer on the bottom of the home page
+ */
+ 
 footer {
   background-color: darken($color-tertiary, 3%);
   font-size: 12px;
