+++ src/stylesheets/components/_navbar.scss
@@ -1,3 +1,6 @@
+/**
+ * Collapsible NavBar on the top of the page
+ */
 .main-navbar {
   height: 56px;
   position: sticky;
