+++ src/stylesheets/components/_sidebar.scss
@@ -1,3 +1,7 @@
+/**
+ * Collapsible sidebars
+ */
+
 .sidebar-page {
   flex-wrap: nowrap;
   height: 100%;
@@ -62,6 +66,9 @@
   top: 0;
 }
 
+/**
+ * The physical toggle button and its associated vertical line
+ */
 .sidebar-toggle {
   background-color: darken($color-secondary, 15%);
   color: $white;
