+++ src/stylesheets/components/_splash-screen.scss
@@ -1,3 +1,6 @@
+/**
+ * Splash screen to display when loading
+ */
 .splash-screen {
   align-items: center;
   display: flex;
