+++ src/stylesheets/components/_toggle.scss
@@ -1,3 +1,6 @@
+/**
+ * Redux Form component toggling between two options
+ */
 .toggle-input {
   display: flex;
   flex-direction: row;
