+++ src/stylesheets/pages/_login.scss
@@ -1,3 +1,7 @@
+// -----------------------------------------------------------------------------
+// This file contains styles that are specific to the login page.
+// -----------------------------------------------------------------------------
+
 .login-background {
   align-items: center;
   display: flex;
