+++ src/stylesheets/pages/_meeting.scss
@@ -1,3 +1,7 @@
+// -----------------------------------------------------------------------------
+// This file contains styles that are specific to the meetings page.
+// -----------------------------------------------------------------------------
+
 .stage-container {
   text-align: center;
 }
