+++ src/stylesheets/vendors/_react-widgets.scss
@@ -1,3 +1,8 @@
+// -----------------------------------------------------------------------------
+// This file contains styles that are specific to the React-Widgets package.
+// We modify the styles of its supplied widgets to fit with the rest of the website.
+// -----------------------------------------------------------------------------
+
 .react-toggle {
   margin-right: 8px;
   vertical-align: middle;
